+++ projects/swimlane/ngx-ui/src/lib/components/dialog/dialog-format.enum.ts
@@ -1,0 +1,4 @@
+export enum DialogFormat {
+  Regular = 'regular',
+  Large = 'large'
+}