+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -6,8 +6,11 @@
 - Fix(DropzoneComponent): addressing minor design review feedback
 - Fix(TipComponent): change error color
 - Enhancement: Add `$color-error` css var
+- Feat(LargeFormatDialogContentComponent): add new component as wrapper for Dialog content
 
-=======
+## 35.3.1 (2021-03-29)
+
+- Fix(DropzoneComponent): addressing minor design review feedback
 
 ## 35.3.0 (2021-03-23)